速度のfloat化

+++ AppleGame/apple.h
@@ -9,8 +9,8 @@ struct APPLE {
 	int flg;		//使用フラグ
 	int type;		//タイプ
 	int img;		//画像
-	int x, y, w, h;	//座標、幅、高さ
-	int speed;		//移動速度
+	float x, y, w, h;	//座標、幅、高さ
+	float speed;		//移動速度
 	int score;		//スコア
 };
